+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Docker Utility Scripts Change Log
 
-(01/17/2022) 1.0.7:
+#### (01/19/2022) 1.1.0:
+- Added functions to parse a 'version' tag to get the Major.Minor.Build-Stage Values
+- Updated Build and Publish functions to utilize the 'new' version tag.
+
+#### (01/17/2022) 1.0.7:
 - Initial Release